requirements.txt
crewai
crewai[tools]
google-genai
pandas
matplotlib
scikit-learn
python-dotenv
seaborn